requirements.txt
pandas>=2.0.0
numpy>=1.24.0
TA-Lib>=0.4.28
asyncio>=3.4.3
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
pydantic>=2.5.0
prometheus-client>=0.19.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
sortedcontainers>=2.4.0
scipy>=1.11.0
pyyaml>=6.0
python-dateutil>=2.8.2
